Break tapyrus_wallet's undeclared reverse dependency on tapyrus_rpc
wallet/rpcwallet.cpp compiles into tapyrus_rpc (so it can touch
CRPCTable directly to register wallet RPC commands), while
tapyrus_rpc declares a dependency on tapyrus_wallet -- not the other
way. But three separate places in tapyrus_wallet called back up into
rpcwallet.cpp-only symbols, an undeclared cycle that happened to keep
linking only because every real target links both libraries together:

- wallet/init.cpp's WalletInit::RegisterRPC called
  RegisterWalletRPCCommands() directly. Split WalletInit's class
  declaration into a new wallet/walletinit.h and moved just that one
  method into wallet/walletinit_rpc.cpp, which compiles into
  tapyrus_rpc; everything else WalletInit does stays in
  wallet/init.cpp (tapyrus_wallet).

- interfaces/wallet.cpp called IssueToken()/IssueReissuableToken()/
  BurnToken(), which lived in rpcwallet.cpp despite taking
  CWallet*/CAmount/CCoinControl& (not JSONRPCRequest) and needing
  nothing RPC-specific. Moved all three into a new
  wallet/tokenissuance.cpp (tapyrus_wallet); rpcwallet.cpp's
  issuetoken/reissuetoken RPC handlers call them exactly as before,
  just now reaching down into tapyrus_wallet instead of sideways
  within their own file.

wallet/rpcdump.cpp (rpcdump.cpp's handlers are pure RPC-command
content, extern-declared and registered by rpcwallet.cpp's own
command table -- no code changes needed there, just its CMake target
membership, in a later commit) rounds out the third call site.
