+++ src/wallet/init.cpp
@@ -11,43 +11,9 @@
 #include <util.h>
 #include <utilmoneystr.h>
 #include <validation.h>
-#include <walletinitinterface.h>
-#include <wallet/rpcwallet.h>
+#include <wallet/walletinit.h>
 #include <wallet/wallet.h>
 #include <wallet/walletutil.h>
-
-class WalletInit : public WalletInitInterface {
-public:
-
-    //! Return the wallets help message.
-    void AddWalletOptions() const override;
-
-    //! Wallets parameter interaction
-    bool ParameterInteraction() const override;
-
-    //! Register wallet RPCs.
-    void RegisterRPC(CRPCTable &tableRPC) const override;
-
-    //! Responsible for reading and validating the -wallet arguments and verifying the wallet database.
-    //  This function will perform salvage on the wallet if requested, as long as only one wallet is
-    //  being loaded (WalletParameterInteraction forbids -salvagewallet, -zapwallettxes or -upgradewallet with multiwallet).
-    bool Verify() const override;
-
-    //! Load wallet databases.
-    bool Open() const override;
-
-    //! Complete startup of wallets.
-    void Start(CScheduler& scheduler) const override;
-
-    //! Flush all wallets in preparation for shutdown.
-    void Flush() const override;
-
-    //! Stop all wallets. Wallets will be flushed first.
-    void Stop() const override;
-
-    //! Close all wallets.
-    void Close() const override;
-};
 
 const WalletInitInterface& g_wallet_init_interface = WalletInit();
 
@@ -160,14 +126,8 @@ bool WalletInit::ParameterInteraction() const
     return true;
 }
 
-void WalletInit::RegisterRPC(CRPCTable &t) const
-{
-    if (gArgs.GetBoolArg("-disablewallet", DEFAULT_DISABLE_WALLET)) {
-        return;
-    }
-
-    RegisterWalletRPCCommands(t);
-}
+// WalletInit::RegisterRPC() is defined in wallet/walletinit_rpc.cpp (tapyrus_rpc),
+// not here -- see the comment on the class declaration in wallet/walletinit.h.
 
 bool WalletInit::Verify() const
 {
